backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-credit-cards==0.4.1
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
openai==1.76.0
packaging==25.0
phonenumbers==9.0.3
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
setuptools==79.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
swagger-spec-validator==3.0.4
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
